% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogluon.R
\name{autogluon_pred}
\alias{autogluon_pred}
\title{Function to predict the GEP-subgroups using Autogluon model}
\usage{
autogluon_pred(x = NULL, exp_type = "log2tpm", use_models = 80)
}
\arguments{
\item{x}{Normalized gene expression matrix or its file path}

\item{exp_type}{deseq2 or log2tpm}

\item{use_models}{models were used}
}
\description{
Function to predict the GEP-subgroups using Autogluon model
}
\examples{
x <- system.file("extdata", "tcga_log2tpm_test.csv", package = "clinaml")
res <- autogluon_pred(x, exp_type = "log2tpm")
x2 <- system.file("extdata", "tcga_deseq2_test.csv", package = "clinaml")
res2 <- autogluon_pred(x2, exp_type = "deseq2")
}
