% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_gep_dat}
\alias{check_gep_dat}
\title{Function to check whether the input fits all features for
GEP-defined subgroups prediction}
\usage{
check_gep_dat(x)
}
\arguments{
\item{x}{Normalized gene expression matrix,
the 1st column is gene id, ensemble id, or gene symbol}
}
\value{
logical
}
\description{
Function to check whether the input fits all features for
GEP-defined subgroups prediction
}
\examples{
check_gep_dat(NULL)
}
