% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinaml.R
\docType{package}
\name{clinaml}
\alias{clinaml}
\title{The 'clinaml' package provides R functions to predict the
gene expression profiling (GEP)-defined subgroups and the GEP-plus risk scores.}
\seealso{
Useful links:

\url{https://github.com/clindet/clinaml}

Report bugs at \url{https://github.com/clindet/clinaml/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
