% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_count.R
\name{convert_featurecounts}
\alias{convert_featurecounts}
\title{Function to convert RNA-Seq reads to gene expression matrix (featureCounts)}
\usage{
convert_featurecounts(
  countfiles = NULL,
  samples = NULL,
  indir = NULL,
  meanFragmentLength = NULL,
  out_raw_counts = "",
  outprefix = NULL,
  gene2symbol = "",
  workers = detectCores()
)
}
\arguments{
\item{samples}{If count_files is NULL, samples was used to
set the value with indir}

\item{indir}{Input salmon counts dir}

\item{meanFragmentLength}{Character vector for sequencing library
length of samples (e.g. 50, 75, 150, 200)}

\item{out_raw_counts}{Output path of raw counts table with extra fields}

\item{outprefix}{Output prefix}

\item{gene2symbol}{Two columns file contains GENEID (e.g. ENSG00000000003.15) and
SYMBOL (TSPAN6)}

\item{workers}{Number of thread for DESeq2 process}

\item{count_files}{Character vector for salmon output count files}
}
\description{
Function to convert RNA-Seq reads to gene expression matrix (featureCounts)
}
\examples{

samples <- readLines("samples")
countfiles <- file.path("../featureCounts", paste0(samples, ".txt"))
out_raw_counts <- "featurecounts/raw_counts_584.txt"
meanFragmentLength <- rep(150, length(samples))
meanFragmentLength[366:477] <- 75
gene2symbol <- fread('~/env/genome/gencode/hg38/v34/tximport.geneid2symbol',
    header = T, data.table = FALSE)
convert_featurecounts(countfiles,
  meanFragmentLength = meanFragmentLength,
  out_raw_counts = out_raw_counts,
  outprefix = "featurecounts/aml-",
  gene2symbol = gene2symbol)
}
