% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_count.R
\name{convert_salmon}
\alias{convert_salmon}
\title{Function to convert RNA-Seq reads to gene expression matrix (Salmon)}
\usage{
convert_salmon(
  count_files = NULL,
  samples = NULL,
  indir = NULL,
  outprefix = NULL,
  txgene_fn = "",
  gene2symbol = "",
  workers = detectCores(),
  ...
)
}
\arguments{
\item{count_files}{Character vector for salmon output count files}

\item{samples}{If count_files is NULL, samples was used to
set the value with indir}

\item{indir}{Input salmon counts dir}

\item{outprefix}{Output prefix}

\item{txgene_fn}{Two columns file contains TXNAME (e.g. ENST00000373020.9) and
GENEID (ENSG00000000003.15)}

\item{gene2symbol}{Two columns file contains GENEID (e.g. ENSG00000000003.15) and
SYMBOL (TSPAN6)}

\item{workers}{Number of thread for DESeq2 process}

\item{...}{countsFromAbundance ("no", "scaledTPM", "lengthScaledTPM", "dtuScaledTPM")
and ignoreAfterBar (TRUE or FALSE). Default is lengthScaledTPM and TRUE}
}
\description{
Function to convert RNA-Seq reads to gene expression matrix (Salmon)
}
\examples{
\dontrun{
  gene2symbol <- fread('~/env/genome/gencode/hg38/v34/tximport.geneid2symbol',
    header = T, data.table = FALSE)
  txgene_fn <- "~/env/genome/gencode/hg38/v34/tximport.genes"
  tx2gene <- fread(txgene_fn)

  samples <- readLines("samples")
  sampledata <- data.frame(sampleName = samples, condition = rep("disease", length(samples)))
  salmon_dir <- "../salmon/output/salmonGTF/"
  salmon_files <- file.path(salmon_dir, samples, "quant.sf")
  convert_salmon(salmon_files,
    outprefix = "salmon/aml",
    txgene_fn = txgene_fn,
    gene2symbol = gene2symbol
  )
}
}
