% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_model_data}
\alias{download_model_data}
\title{Function to open browser and download model data for
predicting GEP-defined subgroups in AML}
\usage{
download_model_data()
}
\description{
Function to open browser and download model data for
predicting GEP-defined subgroups in AML
}
\examples{
download_model_data()
}
