% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_data}
\alias{get_model_data}
\title{Function to get the path of model data. Root path is determined by
option var clinaml.MODEL_DATA or "~/.clinaml/model"}
\usage{
get_model_data()
}
\description{
Function to get the path of model data. Root path is determined by
option var clinaml.MODEL_DATA or "~/.clinaml/model"
}
\examples{
get_model_data()
}
