% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf.R
\name{gtf2exon_len}
\alias{gtf2exon_len}
\title{Function to convert GTF to exon length}
\usage{
gtf2exon_len(
  gtf = NULL,
  outprefix = "exon",
  use_remote_genecode = FALSE,
  version = 38
)
}
\arguments{
\item{gtf}{GTF path (e.g. GENECODE v34)}
}
\description{
Function to convert GTF to exon length
}
