% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf.R
\name{gtf2tximport}
\alias{gtf2tximport}
\title{Function to convert GTF to tximport.geneid2symbol and tximport.genes}
\usage{
gtf2tximport(
  gtf = NULL,
  outprefix = "tximport",
  use_remote_genecode = FALSE,
  version = 38
)
}
\arguments{
\item{gtf}{GTF path (e.g. GENECODE v34)}
}
\description{
Function to convert GTF to tximport.geneid2symbol and tximport.genes
}
\examples{
\dontrun {
  gtf2tximport("gencode.v38.annotation.gtf")
  gtf2tximport("tximport.genes", use_remote_genecode = TRUE)
}
}
