% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_count.R
\name{gtf2tximport}
\alias{gtf2tximport}
\title{Function to convert GTF to tximport.geneid2symbol and tximport.genes}
\usage{
gtf2tximport(gtf)
}
\arguments{
\item{gtf}{GTF path (e.g. GENECODE v34)}
}
\description{
Function to convert GTF to tximport.geneid2symbol and tximport.genes
}
